Manage tasks into submodules
