+++ tasks/package.py
@@ -1,0 +1,7 @@
+import invoke
+
+
+@invoke.task()
+def pack(ctx):
+    """Build a isolated runnable package.
+    """
